binder: Fix Information disclosure vulnerability
The interaction between the kernel /dev/binder and the usermode
Parcel.cpp means that when a Binder object is passed as
BINDER_TYPE_BINDER or BINDER_TYPE_WEAK_BINDER, a pointer to that
object (in the server process) is leaked to the client process as the
cookie value. This leads to a leak of a heap address in many of the
privileged Binder services, including system_server.
The fix is designed to zero out the Binder pointer and cookie before
sending it to the client process

Bug 1812688

Change-Id: Ie5374c3126e226f783e2d043139f9ba61e383bd9
Signed-off-by: Gagan Grover <ggrover@nvidia.com>
Reviewed-on: http://git-master/r/1253265
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>